+++ install.sh
@@ -8,6 +8,7 @@ sudo install -m 755 workon "$BIN_DEST/workon"
 
 sudo mkdir -p "$LIB_DEST"
 sudo rsync -a --delete lib/ "$LIB_DEST/"
+sudo find "$LIB_DEST" -type f -name "*.sh" -exec chmod 755 {} \;
 
 echo "workon installed to: $BIN_DEST/workon"
 echo "workon libs installed to: $LIB_DEST"